requirements.txt
appdirs==1.4.3
attrs==19.3.0
bleach==3.1.1
certifi==2019.11.28
chardet==3.0.4
cobra-projects==0.1.9a0
coverage==5.0.3
Deprecated==1.2.7
distlib==0.3.0
docutils==0.16
entrypoints==0.3
filelock==3.0.12
gitdb==4.0.2
gitdb2==4.0.2
GitPython==3.0.6
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.3.1
keyring==19.0.1
more-itertools==8.2.0
nose==1.3.7
packaging==20.3
pluggy==0.13.1
prompt-toolkit==1.0.14
py==1.8.1
PyGithub==1.43.6
Pygments==2.6.1
PyInquirer==1.0.3
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
PyYAML==5.1.2
readme-renderer==24.0
regex==2020.2.20
requests==2.23.0
six==1.14.0
smmap==3.0.1
tabulate==0.8.5
toml==0.10.0
tox==3.14.5
urllib3==1.25.8
virtualenv==20.0.10
wcwidth==0.1.8
webencodings==0.5.1
wrapt==1.12.1
zipp==3.1.0
